Initial support for Flutter Web - support iframes
