+++ lib/src/widgets/iframe_unsupported.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_html/html_parser.dart';
+import 'package:flutter_html/src/replaced_element.dart';
+import 'package:flutter_html/style.dart';
+import 'package:webview_flutter/webview_flutter.dart';
+import 'package:html/dom.dart' as dom;
+
+/// [IframeContentElement is a [ReplacedElement] with web content.
+class IframeContentElement extends ReplacedElement {
+  final String src;
+  final double width;
+  final double height;
+  final NavigationDelegate navigationDelegate;
+  final UniqueKey key = UniqueKey();
+
+  IframeContentElement({
+    String name,
+    Style style,
+    this.src,
+    this.width,
+    this.height,
+    dom.Element node,
+    this.navigationDelegate,
+  }) : super(name: name, style: style, node: node);
+
+  @override
+  Widget toWidget(RenderContext context) {
+    return Container(
+      width: width ?? (height ?? 150) * 2,
+      height: height ?? (width ?? 300) / 2,
+      child: Text("Iframes are currently not supported in this environment"),
+    );
+  }
+}
